Run example tests in CI (and fix some tests) (#70)
* Use correct transfers_length

* Add CI job for running test cases of the examples

* Include all example contracts

* Run checks on all examples and require clippy-schema

* Apply clippy suggestions

+++ examples/cis1-wccd/src/lib.rs
@@ -376,7 +376,7 @@ fn contract_transfer<A: HasActions>(
 ) -> ContractResult<A> {
     let mut cursor = ctx.parameter_cursor();
     // Parse the number of transfers.
-    let transfers_length: u8 = cursor.get()?;
+    let transfers_length: u16 = cursor.get()?;
     // Get the sender who invoked this contract function.
     let sender = ctx.sender();
